finally figured the issue with pulling HK data async;
so, I had a problem with fetching data from HealthKit and then returning a result AFTER the fetching has been finished, so that we get an actual result instead of a null value (Double 0.0). I also figured out Swift5 returns, which took a few days, but I finally got it. The test I just done with HR data from today works well and displays on screen. Tested on iOS 15.3.1; iPhone 12.