so, this happened - Swift doesn't allow too long structs ... WTF??!
Swift doesn't allow the use of long structs and expressions because the compiler considers them too "complex". But I figured that I don't need to print all the values to the screen, I just need to print them to the log, check if everything works, and then push them to SimtoonAPI. After that's done, I will remove the printouts. So, after all, no harm done in this particular case, but this idiotic idea of making the compiler care about expression complexity is really gonna complicate things if I continue working with Apple platforms ...